Release v2.18.1 - Correcao de crash ao adicionar itens
- Correcao critica: Resolvido problema que causava o fechamento do aplicativo ao adicionar ou editar itens
- Melhoria na estabilidade: Implementado tratamento robusto de excecoes em todas as camadas
- Logs detalhados: Adicionado sistema de logs completo para facilitar diagnÃ³stico
- Validacao de dados: Implementada validacao adicional antes de inserir itens
- Integridade do banco: Melhorada verificacao da lista padrao

Changes:
- Atualizada versao para 2.18.1
- Adicionado logs detalhados em ViewModel, Repository e DAO
- Melhorada validacao de dados no dialogo de adicao
- Garantida integridade da tabela shopping_lists
- Tratamento robusto de excecoes com feedback ao usuario

+++ app/src/main/java/com/example/minhascompras/data/update/AdvancedUpdateManager.kt
@@ -207,7 +207,7 @@ class AdvancedUpdateManager(private val context: Context) {
                 
                 if (attempts < MAX_UPDATE_ATTEMPTS) {
                     // Pausa entre tentativas
-                    delay(2000 * attempts)
+                    delay((2000 * attempts).toLong())
                 }
             }
